Move common entry code to a header
